Add computing reconstruction error, convert vectors to integers before playing them

+++ plot_dictionary_data.py
@@ -25,7 +25,7 @@ def plot_dictionary_data(sampling_rate=48000, call_types=CallType):
       plt.title(call_type.filename)
       button_axes = plt.axes([0.81, 0.05, 0.1, 0.075])
       button = Button(button_axes, 'Play')
-      button.on_clicked(lambda _: sd.play(vec, sampling_rate))
+      button.on_clicked(lambda _: sd.play(vec.astype(np.int16), sampling_rate))
       plt.show()
